+++ tests/ExchangeRateProvider/PDOProviderTest.php
@@ -8,6 +8,7 @@ use Brick\Money\Exception\CurrencyConversionException;
 use Brick\Money\ExchangeRateProvider\PDOProvider;
 use Brick\Money\ExchangeRateProvider\PDOProviderConfiguration;
 use Brick\Money\Tests\AbstractTestCase;
+use Closure;
 
 /**
  * Tests for class PDOProvider.
@@ -18,39 +19,50 @@ class PDOProviderTest extends AbstractTestCase
 {
     /**
      * @dataProvider providerConstructorWithInvalidConfiguration
+     *
+     * @param Closure(): PDOProviderConfiguration $getConfiguration
      */
-    public function testConstructorWithInvalidConfiguration(PDOProviderConfiguration $configuration, string $exceptionMessage) : void
+    public function testConfigurationConstructorThrows(Closure $getConfiguration, string $exceptionMessage) : void
     {
         $this->expectException(\InvalidArgumentException::class);
         $this->expectExceptionMessage($exceptionMessage);
 
-        $pdo = new \PDO('sqlite::memory:');
-
-        new PDOProvider($pdo, $configuration);
+        $getConfiguration();
     }
 
     public function providerConstructorWithInvalidConfiguration() : array
     {
-        $noTableName = new PDOProviderConfiguration();
-
-        $noExchangeRateColumnName = clone $noTableName;
-        $noExchangeRateColumnName->tableName = 'exchange_rate';
-
-        $noSourceCurrency = clone $noExchangeRateColumnName;
-        $noSourceCurrency->exchangeRateColumnName = 'exchange_rate';
-
-        $noTargetCurrency = clone $noSourceCurrency;
-        $noTargetCurrency->sourceCurrencyCode = 'EUR';
-
-        $fixedSourceAndTargetCurrency = clone $noTargetCurrency;
-        $fixedSourceAndTargetCurrency->targetCurrencyCode = 'USD';
-
         return [
-            [$noTableName, '$tableName'],
-            [$noExchangeRateColumnName, '$exchangeRateColumnName'],
-            [$noSourceCurrency, '$sourceCurrencyCode or $sourceCurrencyColumnName'],
-            [$noTargetCurrency, '$targetCurrencyCode or $targetCurrencyColumnName'],
-            [$fixedSourceAndTargetCurrency, '$sourceCurrencyCode and $targetCurrencyCode']
+            [fn () => new PDOProviderConfiguration(
+                tableName: 'exchange_rate',
+                exchangeRateColumnName: 'exchange_rate',
+                targetCurrencyCode: 'EUR',
+            ), 'Invalid configuration: one of $sourceCurrencyCode or $sourceCurrencyColumnName must be set.'],
+            [fn () => new PDOProviderConfiguration(
+                tableName: 'exchange_rate',
+                exchangeRateColumnName: 'exchange_rate',
+                sourceCurrencyCode: 'EUR',
+                sourceCurrencyColumnName: 'source_currency_code',
+                targetCurrencyCode: 'EUR',
+            ), 'Invalid configuration: $sourceCurrencyCode and $sourceCurrencyColumnName cannot be both set.'],
+            [fn () => new PDOProviderConfiguration(
+                tableName: 'exchange_rate',
+                exchangeRateColumnName: 'exchange_rate',
+                sourceCurrencyCode: 'EUR',
+            ), 'Invalid configuration: one of $targetCurrencyCode or $targetCurrencyColumnName must be set.'],
+            [fn () => new PDOProviderConfiguration(
+                tableName: 'exchange_rate',
+                exchangeRateColumnName: 'exchange_rate',
+                sourceCurrencyCode: 'EUR',
+                targetCurrencyCode: 'EUR',
+                targetCurrencyColumnName: 'target_currency_code',
+            ), 'Invalid configuration: $targetCurrencyCode and $targetCurrencyColumnName cannot be both set.'],
+            [fn () => new PDOProviderConfiguration(
+                tableName: 'exchange_rate',
+                exchangeRateColumnName: 'exchange_rate',
+                sourceCurrencyCode: 'EUR',
+                targetCurrencyCode: 'EUR',
+            ), 'Invalid configuration: $sourceCurrencyCode and $targetCurrencyCode cannot be both set.'],
         ];
     }
 
@@ -79,12 +91,12 @@ class PDOProviderTest extends AbstractTestCase
         $statement->execute(['USD', 'EUR', 0.9]);
         $statement->execute(['USD', 'CAD', 1.2]);
 
-        $configuration = new PDOProviderConfiguration();
-
-        $configuration->tableName                = 'exchange_rate';
-        $configuration->sourceCurrencyColumnName = 'source_currency';
-        $configuration->targetCurrencyColumnName = 'target_currency';
-        $configuration->exchangeRateColumnName   = 'exchange_rate';
+        $configuration = new PDOProviderConfiguration(
+            tableName: 'exchange_rate',
+            exchangeRateColumnName: 'exchange_rate',
+            sourceCurrencyColumnName: 'source_currency',
+            targetCurrencyColumnName: 'target_currency',
+        );
 
         $provider = new PDOProvider($pdo, $configuration);
 
@@ -133,12 +145,12 @@ class PDOProviderTest extends AbstractTestCase
         $statement->execute(['USD', 1.1]);
         $statement->execute(['CAD', 1.2]);
 
-        $configuration = new PDOProviderConfiguration();
-
-        $configuration->tableName                = 'exchange_rate';
-        $configuration->sourceCurrencyCode       = 'EUR';
-        $configuration->targetCurrencyColumnName = 'target_currency';
-        $configuration->exchangeRateColumnName   = 'exchange_rate';
+        $configuration = new PDOProviderConfiguration(
+            tableName: 'exchange_rate',
+            exchangeRateColumnName: 'exchange_rate',
+            sourceCurrencyCode: 'EUR',
+            targetCurrencyColumnName: 'target_currency',
+        );
 
         $provider = new PDOProvider($pdo, $configuration);
 
@@ -187,12 +199,12 @@ class PDOProviderTest extends AbstractTestCase
         $statement->execute(['USD', 0.9]);
         $statement->execute(['CAD', 0.8]);
 
-        $configuration = new PDOProviderConfiguration();
-
-        $configuration->tableName                = 'exchange_rate';
-        $configuration->sourceCurrencyColumnName = 'source_currency';
-        $configuration->targetCurrencyCode       = 'EUR';
-        $configuration->exchangeRateColumnName   = 'exchange_rate';
+        $configuration = new PDOProviderConfiguration(
+            tableName: 'exchange_rate',
+            exchangeRateColumnName: 'exchange_rate',
+            sourceCurrencyColumnName: 'source_currency',
+            targetCurrencyCode: 'EUR',
+        );
 
         $provider = new PDOProvider($pdo, $configuration);
 
@@ -247,13 +259,13 @@ class PDOProviderTest extends AbstractTestCase
         $statement->execute([2017, 9, 'EUR', 'USD', 1.15]);
         $statement->execute([2017, 9, 'EUR', 'CAD', 1.25]);
 
-        $configuration = new PDOProviderConfiguration();
-
-        $configuration->tableName                = 'exchange_rate';
-        $configuration->sourceCurrencyColumnName = 'source_currency';
-        $configuration->targetCurrencyColumnName = 'target_currency';
-        $configuration->exchangeRateColumnName   = 'exchange_rate';
-        $configuration->whereConditions          = 'year = ? AND month = ?';
+        $configuration = new PDOProviderConfiguration(
+            tableName: 'exchange_rate',
+            exchangeRateColumnName: 'exchange_rate',
+            sourceCurrencyColumnName: 'source_currency',
+            targetCurrencyColumnName: 'target_currency',
+            whereConditions: 'year = ? AND month = ?',
+        );
 
         $provider = new PDOProvider($pdo, $configuration);
         $provider->setParameters(...$parameters);